Move the embed footer into the app package
The CZ copy of the EmbedFooter becomes
packages/app/src/components/embed-footer.tsx and the per-app barrel,
now empty, is deleted. Unlike the layout, the three copies were not
identical — each hardcoded its own app title and privacy label, which
the next commit turns into messages.
