--- apps/www.volebnikalkulacka.cz/calculator/components/server/components/index.ts
@@ -1,2 +1,0 @@
-export * from "./embed-footer";
-export * from "./layout";
