+++ packages/app/src/components/embed-footer.tsx
@@ -1,0 +1,30 @@
+import { useTranslations } from "next-intl";
+
+import { EmbedAttribution } from "@/components/embed-attribution";
+
+const HEIGHT = "h-11";
+const MARGIN_BOTTOM = "mb-11";
+
+export type EmbedFooter = {
+  attribution?: boolean;
+  homepageHref: string;
+  privacyHref?: string;
+};
+
+export function EmbedFooter({ attribution = true, homepageHref, privacyHref }: EmbedFooter) {
+  const t = useTranslations("koa");
+
+  return (
+    <div className="flex items-baseline gap-4">
+      {attribution && <EmbedAttribution href={homepageHref} title={t("appTitle")} />}
+      {privacyHref && (
+        <a href={privacyHref} target="_blank" rel="noopener noreferrer" className="text-xs text-slate-400 hover:text-slate-600 hover:underline">
+          {t("components.embedFooter.privacy")}
+        </a>
+      )}
+    </div>
+  );
+}
+
+EmbedFooter.heightClassNames = HEIGHT;
+EmbedFooter.marginBottomClassNames = MARGIN_BOTTOM;
